% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_pd_1120}
\alias{boston_pd_1120}
\title{WokeWindows citations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 328141 rows and 45 columns.
}
\source{
\url{https://www.wokewindows.org/exports}

\url{https://github.com/nstory/boston_pd_citations}
}
\usage{
boston_pd_1120
}
\description{
WokeWindows citations
}
\keyword{datasets}
