% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpd_officers}
\alias{bpd_officers}
\title{Boston Police Department officers, 2011-2020}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8470 rows and 7 columns.
}
\usage{
bpd_officers
}
\description{
A lookup table for all police officers in the Boston Police Department.
}
\seealso{
Other Boston Police data: 
\code{\link{bpd_offenses_20}},
\code{\link{bpd_stops_1120}},
\code{\link{bpd_stops_summary}}
}
\concept{Boston Police data}
\keyword{datasets}
