% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpd_stops_1120}
\alias{bpd_stops_1120}
\title{Boston Police Department traffic stops, 2011-2020}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 240092 rows and 6 columns.
}
\usage{
bpd_stops_1120
}
\description{
About 240,000 valid traffic stops made by the Boston Police Department
from 2011 to 2020. These data are culled and cleaned from the full database of over
320,000 traffic citations made during these stops.
}
\seealso{
Other Boston Police data: 
\code{\link{bpd_offenses_20}},
\code{\link{bpd_officers}},
\code{\link{bpd_stops_summary}}
}
\concept{Boston Police data}
\keyword{datasets}
