% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpd_stops_summary}
\alias{bpd_stops_summary}
\title{Summary of Boston Police Department traffic stops by officer and location, 2011-2020}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15913 rows and 7 columns.
}
\usage{
bpd_stops_summary
}
\description{
A summary table of traffic stops grouped by officer and location.
These data feed directly into the \link{visualize_weighted_disparity} calculations.
}
\seealso{
Other Boston Police data: 
\code{\link{bpd_offenses_20}},
\code{\link{bpd_officers}},
\code{\link{bpd_stops_1120}}
}
\concept{Boston Police data}
\keyword{datasets}
