% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_disparity.R
\name{compare_officer_citations}
\alias{compare_officer_citations}
\alias{summarize_officer_citations}
\alias{simulate_officer_citations}
\alias{observe_officer}
\alias{p_value_officer}
\alias{visualize_weighted_disparity}
\title{Weighted disparity measures for police officers}
\usage{
compare_officer_citations(
  citations = comvlong::boston_pd_1120,
  my_officer_id = 9047,
  ...
)

summarize_officer_citations(
  citations = comvlong::boston_pd_1120,
  my_officer_id = 9047,
  race_of_interest = "BLACK"
)

simulate_officer_citations(
  citations = comvlong::boston_pd_1120,
  my_officer_id = 9047,
  race_of_interest = "BLACK",
  n_sims = 100
)

observe_officer(
  citations = comvlong::boston_pd_1120,
  my_officer_id = 9047,
  race_of_interest = "BLACK"
)

p_value_officer(data, y_hat)

visualize_weighted_disparity(data, y_hat)
}
\arguments{
\item{citations}{\code{\link{data.frame}} of police citations}

\item{my_officer_id}{identifier for the officer in question}

\item{...}{current ignored}

\item{race_of_interest}{race of defendant. Default is \code{BLACK}}

\item{n_sims}{Number of simulations}

\item{data}{A \code{\link{data.frame}} of draws from the bootstrap distribution}

\item{y_hat}{The observed weighted disparity}
}
\description{
Weighted disparity measures for police officers
}
\examples{
compare_officer_citations(comvlong::boston_pd_1120, 9047)
summarize_officer_citations(comvlong::boston_pd_1120, 9047)
simulate_officer_citations(comvlong::boston_pd_1120, 9047)
observe_officer(comvlong::boston_pd_1120, 9047, race_of_interest = "BLACK")
observe_officer(comvlong::boston_pd_1120, 9047, race_of_interest = "WHITE")
sims <- simulate_officer_citations()
y_hat <- observe_officer()
p_value_officer(sims, y_hat)
sims <- simulate_officer_citations(n_sims = 10000)
y_hat <- observe_officer()
visualize_weighted_disparity(sims, y_hat)
}
