% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{court_codes}
\alias{court_codes}
\title{Court Codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 174 rows and 2 columns.
\describe{
\item{court_code}{Massachusetts court codes. Mutated from original table so that prefixes align with codes in Boston Police dataset.}
\item{court}{The court corresponding to a court code.}
}
}
\source{
\url{https://www.mass.gov/info-details/trial-court-codes-numerical-listing}
}
\usage{
court_codes
}
\description{
Table of Massachusetts state court codes and their corresponding courts.
}
\keyword{datasets}
