% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woke_windows.R
\name{download_bpd_offenses}
\alias{download_bpd_offenses}
\alias{read_bpd_offenses}
\title{Download Boston Police Department data}
\usage{
download_bpd_offenses()

read_bpd_offenses()
}
\description{
Download Boston Police Department data
}
\examples{
\dontrun{
  x <- download_bpd_offenses()
}

}
